Added the file sizes as an extra check in terms of deleting the files. If they file sizes are different they are either different versions of the same file, or different files all together with that just so happen to have the same file name, don't want to be deleting the files that I might still need, especially if they are different versions of the same file.

+++ src/main/java/com/deleteduplicates/deletefiles/DeleteDuplicateFiles.java
@@ -22,7 +22,12 @@ public class DeleteDuplicateFiles {
             // Need to add one to the index each time the loop iterates through to prevent
             // an index out of bounds exception
             if (duplicateFilesList.size() > (i + 1)) {
-                if (duplicateFilesList.get(i).getFileName().equals(duplicateFilesList.get(i + 1).getFileName())) {
+                /* Check the file name(s) to see if they are equal
+                    Then check the file sizes to see if they are equal
+                    The logic behind this is if the file names are equal and
+                    the sizes are equal then they are the same file, not just different versions */
+                if (duplicateFilesList.get(i).getFileName().equals(duplicateFilesList.get(i + 1).getFileName())
+                        && duplicateFilesList.get(i).getDuplicateFile().length() == duplicateFilesList.get(i + 1).getDuplicateFile().length()) {
                     // Delete the duplicate file
                     fileDeleted = duplicateFilesList.get(i + 1).getDuplicateFile().delete();
                     logger.info("File Deleted: " + duplicateFilesList.get(i + 1).getDuplicateFile() + " " + fileDeleted);
